+++ rocks/__init__.py
@@ -1,7 +1,6 @@
 """For space rocks."""
 import os
 import sys
-
 import rich
 
 # rocks modules
@@ -23,7 +22,7 @@ PATH_META = {
 
 # ------
 # Welcome to rocks
-__version__ = "1.4.6"
+__version__ = "1.4.7"
 
 GREETING = fr"""
                | |